service/requirements.txt
google-adk
pyyaml
loguru
